lab 3

+++ MVS2019/RuslanIvanov/lab1/lab1/Bochka.cpp
@@ -33,7 +33,7 @@ float Bochka::getV()
 	return m_V; 
 }
 
-int Bochka::getPercent()
+float Bochka::getPercent()
 {
 	return  m_spirt * 100;
 }
