+++ MVS2019/RuslanIvanov/lab1/lab1/Bochka.h
@@ -11,7 +11,7 @@ public:
 	Bochka(float,float);
 	~Bochka();
 	void Pereliv(Bochka&);
-	int getPercent();
+	float getPercent();
 	float getSpirt();
 	void  setV(float);
 	float getV();
